% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_paired.R
\name{.rcorr_binom_pair}
\alias{.rcorr_binom_pair}
\title{Correlated Binomial pair via a Gaussian copula}
\usage{
.rcorr_binom_pair(B, n1, p1, n2, p2, rho)
}
\description{
Correlated Binomial pair via a Gaussian copula
}
\keyword{internal}
