% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_plots.R
\name{plot_methods}
\alias{plot_methods}
\title{Forest-style comparison of the interval methods}
\usage{
plot_methods(fit)
}
\arguments{
\item{fit}{an \code{anchored_did_fit}.}
}
\description{
Forest-style comparison of the interval methods
}
