% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_plots.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Posterior density of Delta_DiD (Bayesian method)}
\usage{
plot_posterior(fit, level = 0.95)
}
\arguments{
\item{fit}{an \code{anchored_did_fit} that included the bayes method, or a
\code{did_boot}/\code{did_bayes} result carrying a "draws" attribute.}

\item{level}{shaded credible level.}
}
\description{
Posterior density of Delta_DiD (Bayesian method)
}
